Moved interfaces to root namespace
