+++ MbDotNet/Models/Imposters/HttpImposter.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using MbDotNet.Enums;
-using MbDotNet.Interfaces;
 using MbDotNet.Models.Stubs;
 using Newtonsoft.Json;
