+++ MbDotNet/Models/Stubs/StubBase.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using MbDotNet.Interfaces;
 using MbDotNet.Models.Predicates;
 using MbDotNet.Models.Responses;
 using Newtonsoft.Json;
