+++ MbDotNet/MountebankClient.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using MbDotNet.Interfaces;
 using MbDotNet.Enums;
 using MbDotNet.Models;
 using MbDotNet.Models.Imposters;
